allow verifying non-root nodes
this is especially important if there are multiple signatures in a
document, and there isn't one (or one that you trust) that validates
the entire thing

also refactored the three methods of providing the key into
one, to reduce duplication
